fix: move expand args to module
